Add package level godoc and merge perfdata.go files
